requirements.txt
aiocache==0.12.2
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
discord.py==2.4.0
easy-pil==0.3.3
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
Pillow==10.1.0
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.2.2
yarl==1.9.4
